6. Complex types - mixed list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -4,7 +4,10 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+
+
+
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """
     type-annotated function sum_list which takes a list mxd_lst of
     floats as argument and returns their sum as a float.
